requirements.txt
humanize
PyGithub
prettytable
requests
